requirements.txt
torch
torchvision
pillow
opencv-python-headless
pandas
numpy
scikit-learn
matplotlib
